Create python-publish.yml
Theoretically this would tell pypy when there's a new update every time there is a commit.

Looking into what else is needed. I was planning on trying this with a dummy package first. Assuming that works, I was planning doing a merge request with the main project.

At that point you would need to update the secret to something you know for the official version on pip.